MarkovRandomFields

Documentation for MarkovRandomFields.

MarkovRandomFields.Factor

MarkovRandomFields.MCMC.GibbsSampler

MarkovRandomFields.MCMC.MHSampler

MarkovRandomFields.Models.IsingCoupling

MarkovRandomFields.Models.IsingField

MarkovRandomFields.Models.KSATClause

MarkovRandomFields.Models.PottsCoupling

MarkovRandomFields.Models.PottsField

MarkovRandomFields.TabulatedFactor

MarkovRandomFields.UniformFactor

MarkovRandomFields.MCMC.sample_from_variable_biases

MarkovRandomFields.MCMC.save_observables!

MarkovRandomFields.Test.exact_factor_marginals

MarkovRandomFields.Test.exact_lognormalization

MarkovRandomFields.Test.exact_marginals

MarkovRandomFields.Test.exact_prob

MarkovRandomFields.Test.rand_factor

MarkovRandomFields.Test.rand_mrf

MarkovRandomFields.Factor — Type

Factor

An abstract type representing a factor. It should implement the weight(::Factor, x) method to evaluate the factor for input x.

MarkovRandomFields.TabulatedFactor — Type

TabulatedFactor

A type of Factor constructed by specifying the output to any input in a tabular fashion via an array values.

MarkovRandomFields.UniformFactor — Type

UniformFactor

A type of Factor which returns the same value for any input: it behaves as if it wasn't even there.

MarkovRandomFields.Models.IsingCoupling — Type

IsingCoupling

A type of Factor representing a factor in an Ising distribution. It involves $\pm 1$ variables $\boldsymbol{\sigma}_a=\{\sigma_1, \sigma_2, \ldots\}$. The factor evaluates to $\psi(\boldsymbol{\sigma}_a)=e^{\beta J \prod_{i\in a}\sigma_i}$. A particular case is the pairwise interaction where $a=\{i,j\}$ is a pair of vertices involved in an edge $(ij)$.

Fields

βJ: coupling strength.

MarkovRandomFields.Models.IsingField — Type

IsingField

A type of Factor representing a single-variable external field in an Ising distribution. The factor evaluates to $\psi(\sigma_i)=e^{\beta h \sigma_i}$.

Fields

βh: field strength.

MarkovRandomFields.Models.KSATClause — Type

KSATClause

A type of Factor representing a clause in a k-SAT formula. It involves $\{0,1\}$ variables $\boldsymbol{x}_a=\{x_1, x_2, \ldots, x_k\}$. The factor evaluates to $\psi_a(\boldsymbol{x}_a)=1 - \prod_{i\in a}\delta(x_i, J^a_{i})$.

Fields

J: a vector of booleans.

MarkovRandomFields.Models.PottsCoupling — Type

PottsCoupling

A type of Factor representing a factor in a Potts distribution. It involves k discrete variables $\boldsymbol{x}_a$, each taking a discrete value. The factor evaluates to $\psi(\boldsymbol{x}_a)=e^{\beta J(\boldsymbol{x}_a)}$. A particular case is the pairwise interaction where $a=\{i,j\}$ is a pair of vertices involved in an edge $(ij)$.

Fields

βJ: an Array with $k$ indices storing the (log) values of the factor for each input.

MarkovRandomFields.Models.PottsField — Type

PottsField

A type of Factor representing a single-variable external field in a Potts distribution. The factor evaluates to $\psi(x_i)=e^{\beta h(x_i})$.

Fields

βh: a vector storing the (log) values of the factor for each input.

MarkovRandomFields.MCMC.GibbsSampler — Type

A Gibbs sampler

MarkovRandomFields.MCMC.MHSampler — Type

A Metropolis Hastings sampler

MarkovRandomFields.MCMC.sample_from_variable_biases — Method

Sample a random configuration from the variables' priors

MarkovRandomFields.MCMC.save_observables! — Method

" Save observables in a vector. In case of vector-valued observables any copy of the vector must be made within the observable function

MarkovRandomFields.Test.exact_factor_marginals — Method

exact_factor_marginals(model::MarkovRandomField; p_exact = exact_prob(model))

Exhaustively compute marginal distributions for each factor.

MarkovRandomFields.Test.exact_lognormalization — Method

exact_lognormalization(model::MarkovRandomField)

Exhaustively compute the natural logarithm of the normalization constant.

MarkovRandomFields.Test.exact_marginals — Method

exact_marginals(model::MarkovRandomField; p_exact = exact_prob(model))

Exhaustively compute marginal distributions for each variable.

MarkovRandomFields.Test.exact_prob — Method

exact_prob(model::MarkovRandomField; logZ = exact_normalization(model))

Exhaustively compute the probability of each possible configuration of the variables.

MarkovRandomFields.Test.rand_factor — Method

rand_factor([rng,], nstates)

Return a random Factor whose domain is specified by the iterable nstates.

MarkovRandomFields.Test.rand_mrf — Method

rand_mrf([rng], g::AbstractFactorGraph, nstates; variable_biases)

Return a MarkovRandomField with random factors.

nstates is an iterable containing the number of values that can be taken by each variable.